+++ packages/bundle-source/src/types.js
@@ -12,7 +12,7 @@ export {};
  *  BundleSourceSimple &
  *  BundleSourceWithFormat &
  *  BundleSourceWithOptions &
- *  BundleSourceFallback} BundleSource
+ *  BundleSourceGeneral} BundleSource
  */
 
 /**
@@ -38,7 +38,7 @@ export {};
  * @typedef {<T extends ModuleFormat = 'endoZipBase64'>(
  *  startFilename: string,
  *  format: T,
- *  powers?: { read?: ReadFn; canonical?: CanonicalFn },
+ *  powers?: { read?: ReadFn; canonical?: CanonicalFn, externals?: string[] },
  * ) => Promise<BundleSourceResult<T>>} BundleSourceWithFormat
  */
 
@@ -46,7 +46,7 @@ export {};
  * @typedef {<T extends ModuleFormat = 'endoZipBase64'>(
  *  startFilename: string,
  *  bundleOptions: BundleOptions<T>,
- *  powers?: { read?: ReadFn; canonical?: CanonicalFn },
+ *  powers?: { read?: ReadFn; canonical?: CanonicalFn, externals?: string[] },
  * ) => Promise<BundleSourceResult<T>>} BundleSourceWithOptions
  */
 
@@ -54,8 +54,8 @@ export {};
  * @typedef {<T extends ModuleFormat = 'endoZipBase64'>(
  *   startFilename: string,
  *   formatOrOptions?: T | BundleOptions<T>,
- *   powers?: { read?: ReadFn; canonical?: CanonicalFn },
- * ) => Promise<BundleSourceResult<T>>} BundleSourceFallback
+ *   powers?: { read?: ReadFn; canonical?: CanonicalFn, externals?: string[] },
+ * ) => Promise<BundleSourceResult<T>>} BundleSourceGeneral
  */
 
 /**