+++ packages/pass-style/src/typeGuards.js
@@ -1,8 +1,14 @@
 import { passStyleOf } from './passStyleOf.js';
 
 /** @typedef {import('./types.js').Passable} Passable */
-/** @template T @typedef {import('./types.js').CopyArray<T>} CopyArray */
-/** @template T @typedef {import('./types.js').CopyRecord<T>} CopyRecord */
+/**
+ * @template {Passable} [T=Passable]
+ * @typedef {import('./types.js').CopyArray<T>} CopyArray
+ */
+/**
+ * @template {Passable} [T=Passable]
+ * @typedef {import('./types.js').CopyRecord<T>} CopyRecord
+ */
 /** @typedef {import('./types.js').RemotableObject} Remotable */
 
 const { Fail, quote: q } = assert;