+++ packages/patterns/src/utils.js
@@ -167,3 +167,43 @@ export const applyLabelingError = (func, args, label = undefined) => {
   }
 };
 harden(applyLabelingError);
+
+/**
+ * Makes a one-shot iterable iterator from a provided `next` function.
+ *
+ * @template [T=unknown]
+ * @param {() => IteratorResult<T>} next
+ * @returns {IterableIterator<T>}
+ */
+export const makeIterator = next => {
+  const iter = harden({
+    [Symbol.iterator]: () => iter,
+    next,
+  });
+  return iter;
+};
+harden(makeIterator);
+
+/**
+ * A `harden`ing analog of Array.prototype[Symbol.iterator].
+ *
+ * @template [T=unknown]
+ * @param {Array<T>} arr
+ * @returns {IterableIterator<T>}
+ */
+export const makeArrayIterator = arr => {
+  const { length } = arr;
+  let i = 0;
+  return makeIterator(() => {
+    /** @type {T} */
+    let value;
+    if (i < length) {
+      value = arr[i];
+      i += 1;
+      return harden({ done: false, value });
+    }
+    // @ts-expect-error The terminal value doesn't matter
+    return harden({ done: true, value });
+  });
+};
+harden(makeArrayIterator);