+++ packages/nat/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [4.1.31](https://github.com/endojs/endo/compare/@endo/nat@4.1.30...@endo/nat@4.1.31) (2023-09-12)
+
+**Note:** Version bump only for package @endo/nat
+
+
+
+
+
 ### [4.1.30](https://github.com/endojs/endo/compare/@endo/nat@4.1.28...@endo/nat@4.1.30) (2023-08-07)
 
 **Note:** Version bump only for package @endo/nat